move parameter operator logic into own utils file
