+++ frontend/src/metabase/parameters/components/ParameterValueWidget.jsx
@@ -23,10 +23,8 @@ import {
   makeGetMergedParameterFieldValues,
 } from "metabase/selectors/metadata";
 
-import {
-  getParameterIconName,
-  deriveFieldOperatorFromParameter,
-} from "metabase/meta/Parameter";
+import { getParameterIconName } from "metabase/meta/Parameter";
+import { deriveFieldOperatorFromParameter } from "metabase/parameters/utils/operators";
 import { isDashboardParameterWithoutMapping } from "metabase/meta/Dashboard";
 
 import S from "./ParameterWidget.css";
